moved project page to app/project folder space for consistency
