--- app/(public)/work/ar-story-teller/automatic-seater-assignments.module.scss
@@ -1,44 +1,0 @@
-@import "@/styles/variables";
-@import "@/styles/mixins";
-
-.page {
-  min-height: 100vh;
-  width: 100%;
-  background-color: var(--page-canvas);
-  padding: 3.5rem 0 5rem;
-}
-
-.container {
-  @include global-container;
-}
-
-.hero {
-  text-align: left;
-  margin-bottom: 3rem;
-}
-
-.title {
-  font-family: var(--font-poppins);
-  font-size: clamp(2rem, 4vw, 2.8rem);
-  font-weight: 700;
-  color: #074c5f;
-  margin-bottom: 0.75rem;
-}
-
-.kicker {
-  font-family: var(--font-figtree);
-  font-size: 0.9rem;
-  font-weight: 600;
-  letter-spacing: 0.04em;
-  color: rgba(7, 76, 95, 0.7);
-  margin-bottom: 0.5rem;
-}
-
-.subtitle {
-  font-family: var(--font-figtree);
-  font-size: 1.05rem;
-  line-height: 1.6;
-  color: rgba(1, 17, 20, 0.8);
-  max-width: 720px;
-}
-
