+++ app/projects/automatic-seater-assignments/components/GatedImage.tsx
