+++ app/projects/automatic-seater-assignments/components/OverviewSection.tsx
@@ -4,8 +4,7 @@ import {
   CASE_STUDY_CONTAINER_MAX_WIDTH,
   caseStudyContainerSx,
   caseStudySectionGutterSx,
-} from "../caseStudyLayout";
-import { aosFadeUp } from "../aosProps";
+} from "../layoutConfig";
 
 export type OverviewSectionData = {
   title: string;
@@ -23,7 +22,6 @@ export const OverviewSection = ({ data }: Props) => {
   return (
     <Box
       component="section"
-      {...aosFadeUp()}
       sx={{
         ...caseStudySectionGutterSx,
         background,
