+++ app/projects/automatic-seater-assignments/components/ResearchMethod.tsx
@@ -2,18 +2,16 @@ import React from "react";
 import { Box, Stack, Typography } from "@mui/material";
 
 import { breakpointPx } from "@/lib/responsive/breakpoints";
-import { caseStudySectionGutterSx } from "../caseStudyLayout";
+import { caseStudySectionGutterSx } from "../layoutConfig";
 import type { ResearchMethodBlockData } from "../researchMethodTypes";
 import { ResearchMethodCardShell } from "./ResearchMethodCardShell";
-import { aosFadeUp } from "../aosProps";
 
 type Props = {
   /** One item from `automaticSeaterAssignmentsDataProject.researchMethods` (static today; Firestore later). */
   data: ResearchMethodBlockData;
-  aosDelay?: number;
 };
 
-export const ResearchMethod = ({ data, aosDelay = 0 }: Props) => {
+export const ResearchMethod = ({ data }: Props) => {
   const {
     kicker,
     title,
@@ -114,7 +112,6 @@ export const ResearchMethod = ({ data, aosDelay = 0 }: Props) => {
   return (
     <Box sx={caseStudySectionGutterSx}>
       <Box
-        {...aosFadeUp({ delay: aosDelay, duration: 900 })}
         sx={{
           width: "100%",
           maxWidth: { xs: 655, md: 815, lg: 960, xl: 960 },
