+++ app/projects/automatic-seater-assignments/components/ResearchMethodBlockRenderer.tsx
