+++ app/projects/automatic-seater-assignments/components/ResearchMethodCardShell.tsx
