+++ app/projects/automatic-seater-assignments/components/ResearchMethodImageBlock.tsx
@@ -6,7 +6,6 @@ import { SlideshowLightbox } from "lightbox.js-react";
 import GatedImage from "@/lib/media/GatedImage";
 import { useSignedMediaUrl } from "@/lib/media/useSignedMediaUrl";
 import type { ResearchCardContentBlock } from "../researchMethodTypes";
-import { aosFadeUp } from "../aosProps";
 
 export type ResearchMethodImageBlockData = Extract<
   ResearchCardContentBlock,
@@ -154,7 +153,7 @@ export function ResearchMethodImageBlock({ block }: Props) {
       block.lightboxModalBackground ?? DEFAULT_LIGHTBOX_MODAL_BG;
 
     return (
-      <Stack spacing={1} px={2} {...aosFadeUp({ duration: 1100 })}>
+      <Stack spacing={1} px={2}>
         {blockTitle}
         <SlideshowLightbox
           framework="next"
@@ -189,7 +188,7 @@ export function ResearchMethodImageBlock({ block }: Props) {
   }
 
   return (
-    <Stack spacing={1} px={2} {...aosFadeUp({ duration: 1100 })}>
+    <Stack spacing={1} px={2}>
       {blockTitle}
       <Box
         sx={{
