+++ app/projects/automatic-seater-assignments/components/ReusableComponent.tsx
