+++ app/projects/automatic-seater-assignments/lib/automatic-seater-assignments.firestore.ts
