+++ app/projects/automatic-seater-assignments/researchMethodTypes.ts
